+++ example/namedtracer/foo/foo.go
@@ -18,7 +18,7 @@ var lemonsKey = attribute.Key("ex.com/lemons")
 func SubOperation(ctx context.Context) error {
 	// Using global provider. Alternative is to have application provide a getter
 	// for its component to get the instance of the provider.
-	tr := otel.Tracer("example/namedtracer/foo")
+	tr := otel.Tracer("go.opentelemetry.io/otel/example/namedtracer/foo")
 
 	var span trace.Span
 	_, span = tr.Start(ctx, "Sub operation...")